Задание №7. Каждый робот - это отделный класс. 1-й завод производит 27 вариантов разных роботов. 2-й завод производит 27 вариантов разных роботов.
