+++ src/JAVA_GLk_JC1_29_22/HomeTasks/task8/factory/BodyProd.java
@@ -1,0 +1,9 @@
+package JAVA_GLk_JC1_29_22.HomeTasks.task8.factory;
+
+import JAVA_GLk_JC1_29_22.HomeTasks.task8.bodies.Body;
+import JAVA_GLk_JC1_29_22.HomeTasks.task8.heads.Head;
+
+public interface BodyProd {
+
+    public  Body[] createBodies();
+}
