+++ src/JAVA_GLk_JC1_29_22/HomeTasks/task8/factory/Factory2.java
@@ -4,9 +4,7 @@ import JAVA_GLk_JC1_29_22.HomeTasks.task8.bodies.Body;
 import JAVA_GLk_JC1_29_22.HomeTasks.task8.heads.*;
 import JAVA_GLk_JC1_29_22.HomeTasks.task8.robots.*;
 
-public class Factory2 extends Factory1 {
-
-    protected FactoryType factoryType;
+public class Factory2 extends Factory implements HeadProd {
 
     public  Head[] createHeads() {
