+++ src/JAVA_GLk_JC1_29_22/HomeTasks/task8/factory/HeadProd.java
@@ -1,0 +1,8 @@
+package JAVA_GLk_JC1_29_22.HomeTasks.task8.factory;
+
+import JAVA_GLk_JC1_29_22.HomeTasks.task8.heads.Head;
+
+public interface HeadProd {
+
+    public  Head[] createHeads();
+}
